Renaming create task lib files.
